+++ scalewiz/components/handler_view_devices.py
