+++ scalewiz/components/project_editor.py
