+++ scalewiz/components/project_editor_info.py
