+++ scalewiz/components/project_editor_report.py
