+++ scalewiz/components/scalewiz_log_window.py
