+++ scalewiz/components/scalewiz_main_frame.py
